General fixes.
Other than working on general fixes and advancing a bit more in the
testing, there hasn't been a lot more of news. However by looking at all
the Element tests and doing some debugging, I've found that the
comprehension for the Element tree does not actually return anything
(and so making everything related to elements fail). Hopefully this will
be fixed in a version soon.